requirements.txt
python-binance==1.0.19
pandas==2.1.4
numpy==1.26.2
ta==0.11.0
scipy==1.11.4
aiohttp==3.9.1
websockets==12.0
python-dotenv==1.0.0
openai==1.6.0
loguru==0.7.2
apscheduler==3.10.4
requests==2.31.0
